requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==5.0.7
faiss-cpu==1.6.3
filelock==3.0.12
idna==2.10
importlib-metadata==4.0.1
isodate==0.6.0
joblib==1.0.1
memoized-property==1.0.3
networkx==2.2
nltk==3.6.2
numpy==1.20.2
overrides==3.1.0
packaging==20.9
pandas==1.1.5
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
rdflib==4.2.2
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
scikit-learn==0.24.1
scipy==1.6.3
sentence-transformers==1.1.0
sentencepiece==0.1.95
six==1.15.0
soupsieve==2.2.1
threadpoolctl==2.1.0
tokenizers==0.10.2
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
transformers==4.5.1
typing==3.6.2
typing-extensions==3.7.4.3
urllib3==1.26.4
validators==0.18.2
zipp==3.4.1
